Removed compile warning with constant use of int in place of uint. However this is a breaking change as it seems to cause segfaults.

+++ cigram/lfr_model/cast.cpp
@@ -47,9 +47,9 @@ bool cast_string_to_double (string &b, double &h) {
 	}
 	
 	
-	j=digits_before+1;
+	j = digits_before+1;
 	
-	while (j<b.size()) {
+	while (j < (int) b.size()) {
 		
 		int number=(int(b[j])-48);
 		h+=number*pow(10, digits_before-j);
